Added card class.
The code is pretty self-explainitory except for the index attribute. My plan for the neural network is to take in every known card as input, since this will obviously affect which strategy is best for the bot. To do that, I am going to give each bot an array of zeros with length 40 (which is the number of cards in an Italian deck). Whenever a card is played, the element associated with the index is turned into a 1. Once all the cards have been played, the array will be entirely ones.

--Matteo Raso