Added csv file containing cards.
Instead of trying to save 40 objects in a file, I'm just going to save the attributes as a csv file and read it in when the deck is built.

The columns are, in order, suit, value, points, and index. I'm using the Italian names for the suits and values, which is why it might seem a bit unfamiliar.

--Matteo Raso