Added player class
Contains all the methods needed for the policy and value
updates. It also loads from probability.npy when _init_()
is ran, so it's important that both of the files are kept
in the same directory. It also contains the play_card method,
which will be used to verify its playing ability and to
create an interface for a person to play the bot.