Added a function to make bots play each other
This function will be used to have bots with different
gamma values fight and see which gamma value is better.
Strictly speaking, we don't have to do this, since we
could just see which bot has the highest weighted value,
but this helps make sure we didn't mess up somewhere
when getting the probability or training the bot.