Added decision function
This function takes the output of a neural network, the hand that network
was dealt, and the current briscola, usint the output to decide an action.
If the action is impossible, a different action is chosen instead. This will
definitely cause the network some problems in the beginning, but it should
learn to only play possible moves eventually.

I also included a test function to make sure decision.py works correctly.
Even though the test function is placed in tests/, it needs to be in the
same directory as decision.py to work.