Added method to clear non-input nodes
While playing, the AI has to recalculate the values of the nodes
every time one of the input nodes changes. Clearing the nodes prevents
any miscalculations or the AI accidentally deciding that a node doesn't
need to call get_value. This commit is just a stop-gap measure to fix the
issue, since my main priority right now is to get the AI working, but
it shouldn't be a permanent solution.