BUG: Added failsafe for if a connection gene not in connection_gene[1].incoming is given as input
This doesn't happen very often at all. I would estimate that it happens to about 1 / 400 organisms
while initializing the first species. The error probably happens because of floating point errors
for the weights, but I haven't been able to confirm yet. Definitely not something to be kept in
the AI for the long-term, but the main focus right now is to get a working AI and working on
optimization later on.