NEAT uses less organisms and games played if there's only 1 species
The time it takes to evaluate the fitness for an entire species is
O(n^2 * m), where n is the number of members and m is the num_games.
Since the total number of organisms in a generation is capped, this
means each generation can be calculated much faster if there's more
than 1 species, since if there's only 1 species, n is the total number
of organisms. Since it takes far too long to run the program for the
first few generations, it makes sense to use less organisms and games
played until speciation occurs.