AI now properly detects briscolas
Previously, I accidently made it so that the AI assumes briscola
is always Bastone. After fixing this, it seems that the best AI
favours short term gains, which is not what I expected from
personal experience from playing Briscola. However, we're only
roughly approximating this game as a Markov chain, so it's not
surprising that the AI isn't the best.